pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kirkwood-gpu"
version = "0.1.0"
description = "Realistic-parameter Kirkwood gap formation on a single GPU via CR3BP + symplectic integration"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Dan Ringart" }]
keywords = ["celestial-mechanics", "cr3bp", "symplectic-integrator", "cupy", "gpu-computing", "kirkwood-gaps"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "numpy>=1.24",
    "matplotlib>=3.7",
    "pillow>=10.0",
]

[project.optional-dependencies]
gpu = ["cupy-cuda12x>=13.0"]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]

[project.scripts]
kirkwood-gpu = "kirkwood_gpu.run:main"

[tool.setuptools.packages.find]
include = ["kirkwood_gpu*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
